requirements.txt
snowflake-connector-python>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn==1.2.0
joblib==1.2.0
python-dotenv>=0.19.0